Added an initial_moisture method to crop which merely pulls field_capacity out of its field (for now) shadowing the attribute, which I'll take out later. Took the initial_soil_moisture field out of the running for CropsController (it no longer communicates that attribute to the grid). Removed the column from the grid.
Fixed tests for initial moisture balance when creating a field

Added a pivot fixture for percent cover field use in tests

Changed EtCalculator::linear_interpolation method, added a guard for when the method is called for only one item (e.g., only the emergence date is set for percent cover, so it can't really interpolate anything). Also put in a bunch of extra debug data for if/when the save fails.

Changed the create test to functional/fields_controller to use the same path that the grid does (fields#post_data). Checked that deep drainage is zero for days with pct_cover and ref_et set, nil for surrounding days.

Rejiggered spacing on field_status page to line up the text boxes, and the grid and plot, for a MUCH more pleasing layout.


git-svn-id: https://svn.soils.wisc.edu/svn_repos/IrrigSched/trunk@254 2fdaf364-40d7-4fdf-9344-105978106ad9